Moved AminoConverts back to aminotypes.ts
