+++ packages/stargate/src/aminotypes.ts
@@ -2,7 +2,14 @@
 import { AminoMsg } from "@cosmjs/amino";
 import { EncodeObject } from "@cosmjs/proto-signing";
 
-import { AminoConverter, AminoConverters } from "./aminoconverters";
+export interface AminoConverter {
+  readonly aminoType: string;
+  readonly toAmino: (value: any) => any;
+  readonly fromAmino: (value: any) => any;
+}
+
+/** A map from protobuf type URL to the AminoConverter implementation if supported on chain */
+export type AminoConverters = Record<string, AminoConverter | "not_supported_by_chain">;
 
 function isAminoConverter(
   converter: [string, AminoConverter | "not_supported_by_chain"],
